Delete old bounties folder, replace by current version
