+++ components/bounties/BountyBadge.tsx
