+++ components/databases/focalboard/src/components/kanban/kanbanCard.tsx
@@ -1,7 +1,7 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 import { Box } from '@mui/material'
-import { BountyStatusColours } from 'components/bounties_v2/BountyCard'
+import { BountyStatusColours } from 'components/bounties/BountyCard'
 import { useBounties } from 'hooks/useBounties'
 import millify from "millify"
 import { BOUNTY_LABELS } from 'models'
