+++ pages/[domain]/bounty/[bountyId].tsx
@@ -7,16 +7,16 @@ import Table from '@mui/material/Table';
 import Typography from '@mui/material/Typography';
 import { Bounty } from '@prisma/client';
 import charmClient from 'charmClient';
-import { BountyBadge } from 'components/bounties_v2/BountyBadge';
+import { BountyBadge } from 'components/bounties/BountyBadge';
 import { PageLayout } from 'components/common/page-layout';
-import BountyModal from 'components/bounties_v2/BountyModal';
+import BountyModal from 'components/bounties/BountyModal';
 import CharmEditor from 'components/editor/CharmEditor';
 import { useCurrentSpace } from 'hooks/useCurrentSpace';
 import { useUser } from 'hooks/useUser';
 import { useRouter } from 'next/router';
 import { ReactElement, useEffect, useState, useRef, useCallback } from 'react';
-import { BountyApplicantList } from 'components/bounties_v2/BountyApplicantList';
-import { ApplicationEditorForm } from 'components/bounties_v2/ApplicationEditorForm';
+import { BountyApplicantList } from 'components/bounties/BountyApplicantList';
+import { ApplicationEditorForm } from 'components/bounties/ApplicationEditorForm';
 import { Modal } from 'components/common/Modal';
 import { BountyWithApplications } from 'models';
